pyproject.toml
[tool.poetry]
name = "airflowNASA"
version = "0.1.0"
description = ""
authors = ["Valentin.Kiendl <Valentin.Kiendl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.10.1"
matplotlib = "^3.9.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
